Reducing technical debt
Refactoring to satisfy clang-tidy (great tool)

+++ src/lib/src/libtimeit.cpp
@@ -9,7 +9,7 @@ namespace libtimeit
 
 void init()
 {
-	srand( now() );
+	srand( (unsigned int)now() );
 }
 
 }
